refactor mask gen function, game Id on create
